requirements.txt
imutils==0.5.4
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
